+++ src/Engine/Graphics/Image.cpp
@@ -9,109 +9,6 @@
 
 #include "Engine/Serialization/LegacyImages.h"
 
-#include "Library/Serialization/EnumSerialization.h"
-
-
-struct TextureFrameTable *pTextureFrameTable;
-
-Texture *TextureFrame::GetTexture() {
-    if (!this->tex) {
-        this->tex = assets->getBitmap(this->name);
-    }
-    return this->tex;
-}
-
-void TextureFrameTable::FromFile(const Blob &data_mm6, const Blob &data_mm7, const Blob &data_mm8) {
-    (void) data_mm6;
-    (void) data_mm8;
-
-    deserialize(data_mm7, appendVia<TextureFrame_MM7>(&textures));
-
-    assert(!textures.empty());
-}
-
-void TextureFrameTable::LoadAnimationSequenceAndPalettes(int uFrameID) {
-    if (uFrameID <= textures.size() && uFrameID >= 0) {
-        for (unsigned int i = uFrameID;; ++i) {
-            // this->pTextures[i].uTextureID =
-            // pBitmaps_LOD->LoadTexture(this->pTextures[i].pTextureName,
-            // TEXTURE_DEFAULT);
-
-            // if (this->pTextures[i].uTextureID != -1)
-            //    pBitmaps_LOD->pTextures[this->pTextures[i].uTextureID].palette_id2
-            //    =
-            //    pPaletteManager->LoadPalette(pBitmaps_LOD->pTextures[this->pTextures[i].uTextureID].palette_id1);
-
-            if (textures[i].uFlags & 1) break;
-        }
-    }
-    return;
-}
-
-int64_t TextureFrameTable::FindTextureByName(const std::string &Str2) {
-    std::string name = toLower(Str2);
-
-    for (size_t i = 0; i < textures.size(); ++i)
-        if (textures[i].name == name)
-            return i;
-    return -1;
-}
-
-Texture *TextureFrameTable::GetFrameTexture(int frameId, int time) {
-    int animLength = textures[frameId].uAnimLength;
-
-    if (textures[frameId].uFlags & 1 && animLength != 0) {
-        int step = (time >> 3) % animLength;
-        while (textures[frameId].uAnimTime < step) {
-            step -= textures[frameId].uAnimTime;
-            ++frameId;
-        }
-    }
-
-    return textures[frameId].GetTexture();
-}
-
-int TextureFrameTable::textureFrameAnimLength(int frameID) {
-    int animLength = textures[frameID].uAnimLength;
-    if (textures[frameID].uFlags & 1 && animLength != 0) {
-        return animLength;
-    }
-    return 1;
-}
-
-int TextureFrameTable::textureFrameAnimTime(int frameID) {
-    return textures[frameID].uAnimTime;
-}
-
-void Texture_MM7::Release() {
-    header.pName[0] = 0;
-
-    if (header.pBits & 0x0400) {
-        __debugbreak();
-    }
-
-    if (paletted_pixels != nullptr) {
-        free(paletted_pixels);
-        paletted_pixels = nullptr;
-    }
-
-    if (pPalette24 != nullptr) {
-        free(pPalette24);
-        pPalette24 = nullptr;
-    }
-
-    pLevelOfDetail1 = nullptr;
-
-    memset(&header, 0, sizeof(header));
-}
-
-Texture_MM7::Texture_MM7() {
-    memset(&header, 0, sizeof(header));
-    paletted_pixels = nullptr;
-    pLevelOfDetail1 = nullptr;
-    pPalette24 = nullptr;
-}
-
 GraphicsImage *GraphicsImage::Create(ImageLoader *loader) {
     GraphicsImage *img = new GraphicsImage();
     if (img) {
