+++ src/Engine/Graphics/Indoor.cpp
@@ -15,16 +15,14 @@
 #include "Engine/Graphics/LightsStack.h"
 #include "Engine/Graphics/Outdoor.h"
 #include "Engine/Graphics/Overlays.h"
-#include "Engine/Graphics/PaletteManager.h"
 #include "Engine/Graphics/ParticleEngine.h"
 #include "Engine/Graphics/Texture.h"
+#include "Engine/Graphics/TextureFrameTable.h"
 #include "Engine/Graphics/Sprites.h"
 #include "Engine/Graphics/PortalFunctions.h"
-#include "Engine/Graphics/ClippingFunctions.h"
 #include "Engine/Graphics/Viewport.h"
 #include "Engine/LOD.h"
 #include "Engine/Objects/Actor.h"
-#include "Engine/Objects/Chest.h"
 #include "Engine/Objects/ObjectList.h"
 #include "Engine/Objects/SpriteObject.h"
 #include "Engine/Tables/ItemTable.h"
