+++ src/Engine/Graphics/Texture_MM7.cpp
@@ -1,0 +1,35 @@
+#include "Texture_MM7.h"
+
+#include <cstring>
+#include <cstdlib>
+
+#include "Engine/ErrorHandling.h"
+
+void Texture_MM7::Release() {
+    header.pName[0] = 0;
+
+    if (header.pBits & 0x0400) {
+        __debugbreak();
+    }
+
+    if (paletted_pixels != nullptr) {
+        free(paletted_pixels);
+        paletted_pixels = nullptr;
+    }
+
+    if (pPalette24 != nullptr) {
+        free(pPalette24);
+        pPalette24 = nullptr;
+    }
+
+    pLevelOfDetail1 = nullptr;
+
+    memset(&header, 0, sizeof(header));
+}
+
+Texture_MM7::Texture_MM7() {
+    memset(&header, 0, sizeof(header));
+    paletted_pixels = nullptr;
+    pLevelOfDetail1 = nullptr;
+    pPalette24 = nullptr;
+}
